Tidy things into different places, need to clean up and rename.
