+++ lib/rov/src/ogl/rov_gl_exec.hpp
@@ -2,10 +2,16 @@
 #define ROV_GL_EXEC_INCLUDED_9D909A69_DD1F_40DC_9C6C_6BF5C6914E66
 
 
-namespace ROV_Detail {
+#include <rov/rov.hpp>
+#include "rov_gl.hpp"
+#include "../rov_data.hpp"
 
 
+namespace ROV_Internal {
 
+
+void      ogl_exec(rovGLData *rov_gl_data,
+                   rovData *rov_data);
 
 
 } // ns
