+++ lib/rov/src/rov_data.hpp
@@ -4,64 +4,71 @@
 
 #include <rov/rov.hpp>
 #include <lib/array.hpp>
+#include <vector>
 
 
-namespace ROV_Internal
+namespace ROV_Internal {
+
+
+struct Camera
 {
-  struct Camera
-  {
-    rovMat4 view;
-    rovMat4 proj;
-  };
+  rovMat4 view;
+  rovMat4 proj;
+};
 
-  struct rovMaterial
-  {
-    uint64_t material;
-    size_t draw_calls;
-  };
+struct rovMaterial
+{
+  uint64_t material;
+  size_t draw_calls;
+};
 
-  struct rovDrawCall
-  {
-    uint32_t mesh;
-    rovMat4 world;
-  };
+struct rovDrawCall
+{
+  uint32_t mesh;
+  rovMat4 world;
+};
 
-  struct rovLineDrawCall
-  {
-    rovVec3 start;
-    rovVec3 end;
-    rovVec3 color;
-  };
+struct rovLineDrawCall
+{
+  rovVec3 start;
+  rovVec3 end;
+  rovVec3 color;
+};
 
-  struct rovRenderPass
-  {
-    rovMat4 view;
-    rovMat4 proj;
-    rovViewport viewport;
+struct rovRenderPass
+{
+  rovMat4 view;
+  rovMat4 proj;
+  rovViewport viewport;
 
-    uint32_t clear_flags;
+  uint32_t clear_flags;
 
-    lib::array<rovMaterial> materials;
-    lib::array<rovDrawCall> draw_calls;
-    
-    lib::array<rovLineDrawCall> line_draw_calls;
-  };
+  lib::array<rovMaterial> materials;
+  lib::array<rovDrawCall> draw_calls;
+  
+  lib::array<rovLineDrawCall> line_draw_calls;
+};
 
 
+constexpr uint32_t rov_max_textures = 3;
 
-/*
-  Defined by the API.
-*/
-void
-api_execute(const rovRenderPass passes[], const size_t pass_count);
+struct rovData
+{
+  float       curr_rov_clear_color[4]{0, 0, 0, 1};
+  uint8_t     curr_rov_textures[rov_max_textures]{0};
+  uint32_t    curr_rov_mesh = curr_rov_mesh;
+  uint8_t     curr_rov_mesh_shader = rovShader_Fullbright;
+
+  lib::array<ROV_Internal::Camera>  rov_cameras;
+  std::vector<ROV_Internal::rovRenderPass> rov_render_passes;
+};
 
 
-uint32_t
-api_light_buffer(const rovLight lights[], const size_t light_count);
+uint64_t
+rov_curr_material(rovData *data);
 
 
 }
-
 
 
 #endif // inc guard
